+++ lib/tabulatr/data/row.rb
@@ -1,0 +1,8 @@
+class Tabulatr::Data::Formatting::Row
+
+  attr_accessor :attributes
+
+  def initialize
+    self.attributes = {class: 'tabulatr-row'}
+  end
+end